pyproject.toml
[tool.poetry]
name = "drum-practice-tracker"
version = "0.1.0"
description = ""
authors = ["vlad.shuisky208 <vlad.shuisky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.5"
pillow = "^11.3.0"
psycopg2 = "^2.9.10"
django-environ = "^0.12.0"
python-dotenv = "^1.1.1"
pika = "^1.3.2"
aio-pika = "^9.5.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
